requirements.txt
python-chess>=1.999
numpy>=1.24.0
